refactor(infra): split backend.cuda.intrinsics into codegen/ and cpp/ (#56)
* refactor(infra): extract CUDA codegen infra into tvm.backend.cuda.codegen

Move registry/schema/types/utils/header out of backend.cuda.intrinsics into a
new backend.cuda.codegen package (_schema is de-underscored to schema).
intrinsics/ now holds only content registrations; ptx, tile_primitive, and
tests import the shared machinery from codegen. All FFI global-func names are
unchanged.

codegen/__init__.py imports registry and header for their register_global_func
side effects, which nothing else pulls in anymore. register_backend() now
documents that import_module(".intrinsics") is the sole trigger for device
helper codegen registration, since ptx no longer reaches them transitively.

* refactor(infra): regroup CUDA device helpers into tvm.backend.cuda.cpp

The intrinsics package kept the hardware-domain file split from before the
PTX instructions moved into the ptx table, leaving eight files that each held
a handful of leftovers surrounded by emptied-out scaffolding. Rename it to cpp
-- PTX instructions are intrinsics too, so the old name drew the wrong line --
and regroup by why a helper needs a hand-written C++ body rather than a table
row: builtins (CUDA builtin and library wrappers), asm (spin-wait loops,
barrier pairs, compiler-barrier asm, special-register reads, the legacy
cp.async form), descriptors (wgmma/tcgen05 descriptor encoding and the MMA
validation it enforces), instrument (timers, IKET, printf/trap), and nvshmem.

Drops ~250 lines of dead code and 14 stale scaffold comment blocks left behind
by the table migration: the orphaned _PTX_LD_*/_PTX_ST_* tables and their
parsing helpers in memory.py, two never-registered cluster-barrier builders in
sync.py, and several def-only constants.

The registered codegen set, TAGS, and the rendered header are byte-identical
before and after.

* docs(infra): point stale codegen cross-references at their current homes

Three comments still cited python/tvm/tirx/operator/intrinsics/cuda/, a
directory removed in #19757. Two of them describe code the previous two
commits relocated (the dense instruction descriptor encoder and the
InstrDescriptor bit layout); the third predates the ptx table migration.

* refactor(infra): retire table-migration jargon from cpp helper comments

Section headers still described their contents in the vocabulary of the PTX
form-table audit ("1 form", "Form 2 of mov.type d, sreg") and one header listed
mbarrier and grid-sync helpers that now live elsewhere. Say what each group is
and, where it matters, why it cannot be a table row.

* docs(op): correct the _common modifier-table docstring

It described two consumers: the tvm.tirx.op wrappers, and schema declarations
under tvm.tirx.cuda.intrinsics.* using Choice/IntAttr. That second consumer no
longer exists -- the schema machinery went away with the ptx table migration,
and the package was just renamed. Only the wrappers read these tables now.
