+++ docs/tirx/api/backend.rst
@@ -51,9 +51,17 @@ tvm.backend.cuda.script
    :imported-members:
    :exclude-members: PrimExpr, Op, Call
 
-tvm.backend.cuda.intrinsics
-***************************
-.. automodule:: tvm.backend.cuda.intrinsics
+tvm.backend.cuda.codegen
+************************
+.. automodule:: tvm.backend.cuda.codegen
+   :members:
+   :imported-members:
+   :no-index:
+   :exclude-members: PrimExpr, Op, Call
+
+tvm.backend.cuda.cpp
+********************
+.. automodule:: tvm.backend.cuda.cpp
    :members:
    :imported-members:
    :no-index:
